+++ src/http/routes/index.ts
@@ -16,6 +16,7 @@ import { userItemsRoutes } from './user-items'
 import { usersRoute } from './users'
 import { webhookRoutes } from './webhook'
 import { reviewRepliesRoute } from './review-replies'
+import { socialLinksRoute } from './social-links'
 
 export function routes(app: FastifyInstance) {
   app.register(fastifySwaggerUi, {
@@ -43,6 +44,7 @@ export function routes(app: FastifyInstance) {
   app.register(userItemsRoutes)
   app.register(webhookRoutes)
   app.register(reviewRepliesRoute)
+  app.register(socialLinksRoute)
 
   return
 }