+++ src/http/schemas/users.ts
@@ -86,6 +86,7 @@ export const updateUserBodySchema = z.object({
   bannerPath: z.string().optional(),
   imagePath: z.string().optional(),
   username: z.string().optional(),
+  biography: z.string().optional(),
 })
 
 export const updateUserSchema = {