+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Add CRUD operations for social links. [#37](https://github.com/plotwist-app/plotwist-backend/pull/37)
 - Add CRUD operations for Review Replies. [#28](https://github.com/plotwist-app/plotwist-backend/pull/28)
 
 ### Changed